requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-cors-headers>=4.3.1
PyJWT>=2.6.0
Pillow>=10.0.0